feat: add new schedule table
