+++ client/src/services/course.ts
@@ -70,6 +70,8 @@ export interface CourseEvent {
   organizer: UserBasic;
   detailsUrl: string;
   broadcastUrl: string;
+  special?: string;
+  duration?: number;
 }
 
 export interface CourseUser {
